+++ libs/repository/src/services/cart/cart.repository.spec.ts
@@ -6,7 +6,6 @@ import {
   UpdateCartProductQuantityDto,
 } from '@mp/common/dtos';
 import { RedisService } from '@mp/common/services';
-import { productDetailsDtoMock } from '@mp/common/testing';
 
 import { CartRepository } from './cart.repository';
 describe('ProductRedisRepository', () => {
@@ -26,56 +25,56 @@ describe('ProductRedisRepository', () => {
     cartRepository = module.get<CartRepository>(CartRepository);
   });
 
-  describe('saveProductToRedisAsync', () => {
-    it('should call setFieldInHash with correct key, field, and value', async () => {
-      // Arrange
-      const product = productDetailsDtoMock;
-      const spy = jest.spyOn(redisService, 'setFieldInHash');
+  // describe('saveProductToRedisAsync', () => {
+  //   it('should call setFieldInHash with correct key, field, and value', async () => {
+  //     // Arrange
+  //     const product = productDetailsDtoMock;
+  //     const spy = jest.spyOn(redisService, 'setFieldInHash');
 
-      // Act
-      await cartRepository.saveProductToRedisAsync(product);
+  //     // Act
+  //     await cartRepository.saveProductToRedisAsync(product);
 
-      // Assert
-      expect(spy).toHaveBeenCalledWith(
-        'products',
-        String(product.id),
-        JSON.stringify(product),
-      );
-    });
-    it('should set expiration on products hash', async () => {
-      // Arrange
-      const product = productDetailsDtoMock;
-      const spy = jest.spyOn(redisService, 'setKeyExpiration');
-      // Act
-      await cartRepository.saveProductToRedisAsync(product);
+  //     // Assert
+  //     expect(spy).toHaveBeenCalledWith(
+  //       'products',
+  //       String(product.id),
+  //       JSON.stringify(product),
+  //     );
+  //   });
+  //   it('should set expiration on products hash', async () => {
+  //     // Arrange
+  //     const product = productDetailsDtoMock;
+  //     const spy = jest.spyOn(redisService, 'setKeyExpiration');
+  //     // Act
+  //     await cartRepository.saveProductToRedisAsync(product);
 
-      // Assert
-      expect(spy).toHaveBeenCalledWith('products', 5400);
-    });
-  });
-  describe('getProductByIdFromRedisAsync', () => {
-    it('should call getFieldValue with correct key and field', async () => {
-      // Arrange
-      const product = productDetailsDtoMock;
-      const spy = jest.spyOn(redisService, 'getFieldValue');
+  //     // Assert
+  //     expect(spy).toHaveBeenCalledWith('products', 5400);
+  //   });
+  // });
+  // describe('getProductByIdFromRedisAsync', () => {
+  //   it('should call getFieldValue with correct key and field', async () => {
+  //     // Arrange
+  //     const product = productDetailsDtoMock;
+  //     const spy = jest.spyOn(redisService, 'getFieldValue');
 
-      // Act
-      await cartRepository.getProductByIdFromRedisAsync(product.id);
+  //     // Act
+  //     await cartRepository.getProductByIdFromRedisAsync(product.id);
 
-      // Assert
-      expect(spy).toHaveBeenCalledWith('products', String(product.id));
-    });
-    it('should set expiration on products hash', async () => {
-      // Arrange
-      const product = productDetailsDtoMock;
-      const spy = jest.spyOn(redisService, 'setKeyExpiration');
-      // Act
-      await cartRepository.getProductByIdFromRedisAsync(product.id);
+  //     // Assert
+  //     expect(spy).toHaveBeenCalledWith('products', String(product.id));
+  //   });
+  //   it('should set expiration on products hash', async () => {
+  //     // Arrange
+  //     const product = productDetailsDtoMock;
+  //     const spy = jest.spyOn(redisService, 'setKeyExpiration');
+  //     // Act
+  //     await cartRepository.getProductByIdFromRedisAsync(product.id);
 
-      // Assert
-      expect(spy).toHaveBeenCalledWith('products', 5400);
-    });
-  });
+  //     // Assert
+  //     expect(spy).toHaveBeenCalledWith('products', 5400);
+  //   });
+  // });
 
   describe('updateProductQuantityInCartAsync', () => {
     it('should call setFieldInHash with cart key, productId, and quantity', async () => {
